Conditional path template's expressions are validated. Error raised for: invalid keywords, cross-type comparison. Unit test created

+++ tools/python/boutiques/tests/test_output_files.py
@@ -19,7 +19,8 @@ class TestOutputFiles(TestCase):
         test_desc_path = op.join(
             op.split(bfile)[0],
             "tests/output_files/test_desc.json")
-        os.remove(test_desc_path)
+        if os.path.exists(test_desc_path):
+            os.remove(test_desc_path)
 
     def test_output_conditional_path_template_validity(self):
         base_path = op.join(op.split(bfile)[0], "tests/output_files/")
@@ -147,4 +148,23 @@ class TestOutputFiles(TestCase):
         process_output = validate_wrong_ID.stdout.read()
         self.assertIn("OutputError: \"out2\" contains non-python "
                       "keyword and non-ID string: \"bob\"",
+                      process_output.decode())
+
+    def test_conditional_path_template_comparison_types(self):
+        base_path = op.join(op.split(bfile)[0], "tests/output_files/")
+        test_desc_path = op.join(
+            base_path, "test_cond_output_invalid_types.json")
+
+        command = ("bosh validate " + test_desc_path)
+        validate_wrong_ID = subprocess.Popen(command, shell=True,
+                                             stdout=subprocess.PIPE)
+        process_output = validate_wrong_ID.stdout.read()
+        self.assertIn("\"opt1 and (opt2 > \"10.00\")\" contains "
+                      "invalid conditional expression.",
+                      process_output.decode())
+        self.assertIn("\"opt1 and opt2 and (opt1 == \"False\")\" contains "
+                      "invalid conditional expression.",
+                      process_output.decode())
+        self.assertIn("\"param1 < 10.0\" contains "
+                      "invalid conditional expression.",
                       process_output.decode())
